design-system: add opaque background modifier and button action for PNavigationBar
